fix(rule-pack): use real reference identity for the bundled-default gate
isBundledDefault was a plain field on the RulePack object, so it survived
object spread along with every other own property: a caller who did
`{ ...provisionalRulePack, rules: forgedRules }` kept isBundledDefault: true
while replacing the data the citation gate was supposed to be vouching for.
Verified directly with a spread-and-override reproduction before fixing.

Moves the bundled-default singleton itself into core/default-pack.ts so
core/runner.ts can compare pack === provisionalRulePack by genuine reference
identity, which spread cannot forge because it always allocates a new
object. rule-pack/provisional-pack.ts re-exports the same binding, so every
existing import path is unchanged. No public API signature change.
